refactor: move serializer to a separate subpackage
