+++ tests/test_system.py
@@ -10,9 +10,10 @@ from hypothesis import strategies as st
 from sklearn.linear_model import LinearRegression
 
 from autora.experimentalist.grid import grid_pool
+from autora.serializer import SerializersSupported, load_state
 from autora.state import StandardState, State, estimator_on_state, on_state
 from autora.variable import Variable, VariableCollection
-from autora.workflow.__main__ import SerializersSupported, load_state, main
+from autora.workflow.__main__ import main
 
 _logger = logging.getLogger(__name__)
